Fix compiler warnings across src/
- Fix undefined behavior: unsequenced modifications to curByte in d_net.cpp
- Fix deprecated implicit this capture in lambdas (use explicit [this] or
  [this, ...] captures instead of [=])
- Fix signed/unsigned comparison warnings against MAXPLAYERS and size_t
- Fix format string mismatches (ptrdiff_t vs size_t)
- Fix uint32_t vs int mismatch for particle index in p_effect.cpp

+++ src/playsim/p_acs.cpp
@@ -4588,7 +4588,7 @@ int DLevelScript::GetPlayerInput(int playernum, int inputnum)
 		}
 		p = activator->player;
 	}
-	else if (playernum >= MAXPLAYERS || !Level->PlayerInGame(playernum))
+	else if ((size_t)playernum >= MAXPLAYERS || !Level->PlayerInGame(playernum))
 	{
 		return 0;
 	}
@@ -5524,7 +5524,7 @@ int DLevelScript::CallFunction(int argCount, int funcIndex, int32_t *args, int &
 		case ACSF_GetAirSupply:
 			MIN_ARG_COUNT(1);
 		{
-			if (args[0] < 0 || args[0] >= MAXPLAYERS || !Level->PlayerInGame(args[0]))
+			if (args[0] < 0 || (size_t)args[0] >= MAXPLAYERS || !Level->PlayerInGame(args[0]))
 			{
 				return 0;
 			}
@@ -5537,7 +5537,7 @@ int DLevelScript::CallFunction(int argCount, int funcIndex, int32_t *args, int &
 		case ACSF_SetAirSupply:
 			MIN_ARG_COUNT(2);
 		{
-			if (args[0] < 0 || args[0] >= MAXPLAYERS || !Level->PlayerInGame(args[0]))
+			if (args[0] < 0 || (size_t)args[0] >= MAXPLAYERS || !Level->PlayerInGame(args[0]))
 			{
 				return 0;
 			}
@@ -5560,7 +5560,7 @@ int DLevelScript::CallFunction(int argCount, int funcIndex, int32_t *args, int &
 		case ACSF_GetArmorType:
 			MIN_ARG_COUNT(2);
 		{
-			if (args[1] < 0 || args[1] >= MAXPLAYERS || !Level->PlayerInGame(args[1]))
+			if (args[1] < 0 || (size_t)args[1] >= MAXPLAYERS || !Level->PlayerInGame(args[1]))
 			{
 				return 0;
 			}
@@ -9849,7 +9849,7 @@ scriptwait:
 			break;
 
 		case PCD_PLAYERINGAME:
-			if (STACK(1) < 0 || STACK(1) >= MAXPLAYERS)
+			if (STACK(1) < 0 || (size_t)STACK(1) >= MAXPLAYERS)
 			{
 				STACK(1) = false;
 			}
@@ -9860,7 +9860,7 @@ scriptwait:
 			break;
 
 		case PCD_PLAYERISBOT:
-			if (STACK(1) < 0 || STACK(1) >= MAXPLAYERS || !Level->PlayerInGame(STACK(1)))
+			if (STACK(1) < 0 || (size_t)STACK(1) >= MAXPLAYERS || !Level->PlayerInGame(STACK(1)))
 			{
 				STACK(1) = false;
 			}
@@ -10059,7 +10059,7 @@ scriptwait:
 			break;
 
 		case PCD_PLAYERCLASS:		// [GRB]
-			if (STACK(1) < 0 || STACK(1) >= MAXPLAYERS || !Level->PlayerInGame(STACK(1)))
+			if (STACK(1) < 0 || (size_t)STACK(1) >= MAXPLAYERS || !Level->PlayerInGame(STACK(1)))
 			{
 				STACK(1) = -1;
 			}
@@ -10070,7 +10070,7 @@ scriptwait:
 			break;
 
 		case PCD_GETPLAYERINFO:		// [GRB]
-			if (STACK(2) < 0 || STACK(2) >= MAXPLAYERS || !Level->PlayerInGame(STACK(2)))
+			if (STACK(2) < 0 || (size_t)STACK(2) >= MAXPLAYERS || !Level->PlayerInGame(STACK(2)))
 			{
 				STACK(2) = -1;
 			}
